Moved various tests to other packages
