+++ Tribler/Test/Core/APIImplementation/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains various tests for the classes found in the APIImplementation package.
+"""
