+++ Tribler/Test/Core/APIImplementation/test_launch_many_cores.py
