+++ Tribler/Test/Core/CacheDB/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for the SQLite database used by Tribler.
+"""
