+++ Tribler/Test/Core/CacheDB/test_sqlitecachedbhandler_torrents.py
@@ -1,14 +1,15 @@
-from binascii import unhexlify
 import os
+from binascii import unhexlify
 from shutil import copy as copyfile
+
 from twisted.internet.defer import inlineCallbacks
 
-from Tribler.Core.Category.Category import Category
 from Tribler.Core.CacheDB.SqliteCacheDBHandler import TorrentDBHandler, MyPreferenceDBHandler, ChannelCastDBHandler
 from Tribler.Core.CacheDB.sqlitecachedb import str2bin
+from Tribler.Core.Category.Category import Category
 from Tribler.Core.TorrentDef import TorrentDef
 from Tribler.Core.leveldbstore import LevelDbStore
-from Tribler.Test.Core.test_sqlitecachedbhandler import AbstractDB
+from Tribler.Test.Core.CacheDB.test_sqlitecachedbhandler import AbstractDB
 from Tribler.Test.test_as_server import TESTS_DATA_DIR
 from Tribler.dispersy.util import blocking_call_on_reactor_thread
