+++ Tribler/Test/Core/TorrentChecker/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains various tests for the torrent checker.
+"""
