+++ Tribler/Test/Core/TorrentChecker/test_torrentchecker.py
