+++ Tribler/Test/Core/Utilities/test_encoding.py
