+++ Tribler/Test/Core/Utilities/test_install_dir.py
