+++ Tribler/Test/Core/Utilities/test_instrumentation.py
