+++ Tribler/Test/Core/Utilities/test_win32regchecker.py
